safety-app/backend/livekit-agent/requirements.txt
livekit
livekit-agents[openai]
livekit-plugins-silero
python-dotenv
